requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
DateTime==5.5
idna==3.10
packaging==25.0
pillow==11.2.1
pycparser==2.22
pycryptodome==3.22.0
pyOpenSSL==25.0.0
pyotp==2.9.0
pytz==2025.2
qrcode==8.1
requests==2.32.3
secretsharing==0.2.6
setuptools==80.0.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
zope.interface==7.2
